+++ openslides/utils/consumers.py
@@ -1,5 +1,6 @@
 from typing import Any, Dict, List, Optional
 
+import jsonschema
 from asgiref.sync import sync_to_async
 from channels.db import database_sync_to_async
 from channels.generic.websocket import AsyncJsonWebsocketConsumer
@@ -16,7 +17,73 @@ from .collection import (
 )
 
 
-class SiteConsumer(AsyncJsonWebsocketConsumer):
+class ProtocollAsyncJsonWebsocketConsumer(AsyncJsonWebsocketConsumer):
+    """
+    Mixin for JSONWebsocketConsumers, that speaks the a special protocol.
+    """
+    schema = {
+        "$schema": "http://json-schema.org/draft-04/schema#",
+        "title": "OpenSlidesWebsocketProtocol",
+        "description": "The base packages that OpenSlides sends between the server and the client.",
+        "type": "object",
+        "properties": {
+            "type": {
+                "description": "Defines what kind of packages is packed.",
+                "type": "string",
+                "pattern": "notify",  # The server can sent other types
+            },
+            "content": {
+                "description": "The content of the package.",
+            },
+            "id": {
+                "description": "An identifier of the package.",
+                "type": "string",
+            },
+            "in_response": {
+                "description": "The id of another package that the other part sent before.",
+                "type": "string",
+            }
+        },
+        "required": ["type", "content", "id"],
+    }
+
+    async def send_json(self, type: str, content: Any, id: Optional[str] = None, in_response: Optional[str] = None) -> None:
+        """
+        Sends the data with the type.
+        """
+        out = {'type': type, 'content': content}
+        if id:
+            out['id'] = id
+        if in_response:
+            out['in_response'] = in_response
+        await super().send_json(out)
+
+    async def receive_json(self, content: Any) -> None:
+        """
+        Receives the json data, parses it and calls receive_content.
+        """
+        try:
+            jsonschema.validate(content, self.schema)
+        except jsonschema.ValidationError as err:
+            try:
+                in_response = content['id']
+            except (TypeError, KeyError):
+                # content is not a dict (TypeError) or has not the key id (KeyError)
+                in_response = None
+
+            await self.send_json(
+                type='error',
+                content=str(err),
+                in_response=in_response)
+            return
+
+        await self.receive_content(type=content['type'], content=content['content'], id=content['id'])
+
+    async def receive_content(self, type: str, content: object, id: str) -> None:
+        raise NotImplementedError("ProtocollAsyncJsonWebsocketConsumer needs the method receive_content()")
+
+
+class SiteConsumer(ProtocollAsyncJsonWebsocketConsumer):
     """
     Websocket Consumer for the site.
     """
@@ -36,9 +103,9 @@ class SiteConsumer(AsyncJsonWebsocketConsumer):
         else:
             await self.accept()
             data = await startup_data(self.scope['user'])
-            await self.send_json(data)
+            await self.send_json(type='autoupdate', content=data)
 
-    async def receive_json(self, content: Any) -> None:
+    async def receive_content(self, type: str, content: Any, id: str) -> None:
         """
         If we recieve something from the client we currently just interpret this
         as a notify message.
@@ -47,27 +114,28 @@ class SiteConsumer(AsyncJsonWebsocketConsumer):
         channel name so that a receiver client may reply to the sender or to all
         sender's instances.
         """
-        if notify_message_is_valid(content):
-            await self.channel_layer.group_send(
-                "projector",
-                {
-                    "type": "send_notify",
-                    "incomming": content,
-                    "senderReplyChannelName": self.channel_name,
-                    "senderUserId": self.scope['user'].id or 0,
-                },
-            )
-            await self.channel_layer.group_send(
-                "site",
-                {
-                    "type": "send_notify",
-                    "incomming": content,
-                    "senderReplyChannelName": self.channel_name,
-                    "senderUserId": self.scope['user'].id or 0,
-                },
-            )
-        else:
-            await self.send_json({'error': 'invalid message'})
+        if type == 'notify':
+            if notify_message_is_valid(content):
+                await self.channel_layer.group_send(
+                    "projector",
+                    {
+                        "type": "send_notify",
+                        "incomming": content,
+                        "senderReplyChannelName": self.channel_name,
+                        "senderUserId": self.scope['user'].id or 0,
+                    },
+                )
+                await self.channel_layer.group_send(
+                    "site",
+                    {
+                        "type": "send_notify",
+                        "incomming": content,
+                        "senderReplyChannelName": self.channel_name,
+                        "senderUserId": self.scope['user'].id or 0,
+                    },
+                )
+            else:
+                await self.send_json(type='error', content='Invalid notify message', in_response=id)
 
     async def send_notify(self, event: Dict[str, Any]) -> None:
         """
@@ -89,7 +157,7 @@ class SiteConsumer(AsyncJsonWebsocketConsumer):
                 out.append(item)
 
         if out:
-            await self.send_json(out)
+            await self.send_json(type='notify', content=out)
 
     async def send_data(self, event: Dict[str, Any]) -> None:
         """
@@ -111,10 +179,10 @@ class SiteConsumer(AsyncJsonWebsocketConsumer):
                 collection_string=collection_string,
                 id=id,
                 action='deleted'))
-        await self.send_json(output)
+        await self.send_json(type='autoupdate', content=output)
 
 
-class ProjectorConsumer(AsyncJsonWebsocketConsumer):
+class ProjectorConsumer(ProtocollAsyncJsonWebsocketConsumer):
     """
     Websocket Consumer for the projector.
     """
@@ -131,14 +199,14 @@ class ProjectorConsumer(AsyncJsonWebsocketConsumer):
         await self.accept()
 
         if not await database_sync_to_async(has_perm)(user, 'core.can_see_projector'):
-            await self.send_json({'text': 'No permissions to see this projector.'})
+            await self.send_json(type='error', content='No permissions to see this projector.')
             # TODO: Shouldend we just close the websocket connection with an error message?
             #       self.close(code=4403)
         else:
             out = await sync_to_async(projector_startup_data)(projector_id)
-            await self.send_json(out)
+            await self.send_json(type='autoupdate', content=out)
 
-    async def receive_json(self, content: Any) -> None:
+    async def receive_content(self, type: str, content: Any, id: str) -> None:
         """
         If we recieve something from the client we currently just interpret this
         as a notify message.
@@ -187,7 +255,7 @@ class ProjectorConsumer(AsyncJsonWebsocketConsumer):
                 out.append(item)
 
         if out:
-            await self.send_json(out)
+            await self.send_json(type='notify', content=out)
 
     async def send_data(self, event: Dict[str, Any]) -> None:
         """
@@ -198,7 +266,7 @@ class ProjectorConsumer(AsyncJsonWebsocketConsumer):
 
         output = await projector_sync_send_data(projector_id, collection_elements)
         if output:
-            await self.send_json(output)
+            await self.send_json(type='autoupdate', content=output)
 
 
 async def startup_data(user: Optional[CollectionElement], change_id: int = 0) -> List[Any]:
@@ -287,19 +355,33 @@ def notify_message_is_valid(message: object) -> bool:
     """
     Returns True, when the message is a valid notify_message.
     """
-    if not isinstance(message, list):
-        # message has to be a list
+    schema = {
+        "$schema": "http://json-schema.org/draft-04/schema#",
+        "title": "Notify elements.",
+        "description": "Elements that one client can send to one or many other clients.",
+        "type": "array",
+        "items": {
+            "type": "object",
+            "properties": {
+                "projectors": {
+                    "type": "array",
+                    "items": {"type": "integer"},
+                },
+                "reply_channels": {
+                    "type": "array",
+                    "items": {"type": "string"},
+                },
+                "users": {
+                    "type": "array",
+                    "items": {"type": "integer"},
+                }
+            }
+        },
+        "minItems": 1,
+    }
+    try:
+        jsonschema.validate(message, schema)
+    except jsonschema.ValidationError:
         return False
-
-    if not message:
-        # message must contain at least one element
-        return False
-
-    for element in message:
-        if not isinstance(element, dict):
-            # All elements have to be a dict
-            return False
-    # TODO: There could be more checks. For example 'users' has to be a list of int
-    #       Check could be done with json-schema:
-    #       https://pypi.org/project/jsonschema/
-    return True
+    else:
+        return True